PoC of curriculum learning integration.
